requirements.txt
numpy>=1.22
matplotlib>=3.6
requests>=2.31
Pillow>=9.2
jupyterlab>=4.0
ipykernel>=6.20
matplotlib>=3.6
ipympl>=0.9